+++ src/commonMain/kotlin/csense/kotlin/tests/assertions/Coroutines.kt
@@ -1,4 +1,3 @@
-@file:OptIn(ExperimentalCoroutinesApi::class, ExperimentalCoroutinesApi::class)
 @file:Suppress("unused")
 
 package csense.kotlin.tests.assertions
@@ -8,6 +7,17 @@ import kotlinx.coroutines.test.*
 import kotlin.coroutines.*
 import kotlin.time.*
 
+public fun CoroutineScope.assertDispatcherDefault() {
+    assertDispatcher(Dispatchers.Default)
+}
+
+public fun CoroutineScope.assertDispatcherMain() {
+    assertDispatcher(Dispatchers.Main)
+}
+
+public fun CoroutineScope.assertDispatcherUnconfined() {
+    assertDispatcher(Dispatchers.Unconfined)
+}
 
 public fun CoroutineScope.assertDispatcher(otherDispatcher: CoroutineDispatcher) {
     getCurrentDispatcher().assertAs(otherDispatcher)
@@ -18,7 +28,3 @@ private fun CoroutineScope.getCurrentDispatcher() = coroutineContext.getCurrentD
 private fun CoroutineContext.getCurrentDispatcher(): CoroutineDispatcher =
     this[ContinuationInterceptor] as CoroutineDispatcher
 
-
-public fun TestScope.advanceTimeBy(duration: Duration) {
-    advanceTimeBy(delayTimeMillis = duration.inWholeMilliseconds)
-}