+++ changelog.md
@@ -1,3 +1,15 @@
+# 0.0.60
+- now available on tier 1 & tier 2 (mostly) targets + mingw
+- targets java 11
+- only ships with junit 5 on jvm.
+- added
+    - assertDispatcher<Default, Main, Unconfined, IO>
+    - assertNotByEquals
+    - error(s) to when using assertNull / assertNotNull wrongly (eg on a non-nullable value)
+  
+- changed
+  - assertIs(apply) to accept null, which of cause will fail the test
+  - 
 # 0.0.59
 
 - using kotlin 1.7.20